Рефакторит код и наводит порядок в модулях
